requirements.txt
torchvision~=0.13.1
streamlit~=1.12.2
torch~=1.12.1
Pillow~=9.2.0
tqdm~=4.64.1